+++ sentry-ruby/spec/sentry/vernier/profiler_spec.rb
@@ -2,7 +2,7 @@
 
 require "sentry/vernier/profiler"
 
-RSpec.describe Sentry::Vernier::Profiler, when: { ruby_version?: [:>=, "3.3"], ruby_engine?: "ruby" } do
+RSpec.describe Sentry::Vernier::Profiler, when: :vernier_installed? do
   subject(:profiler) { described_class.new(Sentry.configuration) }
 
   before do
@@ -283,7 +283,7 @@ RSpec.describe Sentry::Vernier::Profiler, when: { ruby_version?: [:>=, "3.3"], r
           expect(thread2[:name]).to eq("thread-bar-1")
         end
 
-        it 'has correct frames', when: { ruby_version?: [:>=, "3.3"], ruby_engine?: "ruby" } do
+        it 'has correct frames' do
           frames = profiler.to_h[:profile][:frames]
 
           foo_frame = frames.find { |f| f[:function] =~ /foo/ }
@@ -296,7 +296,7 @@ RSpec.describe Sentry::Vernier::Profiler, when: { ruby_version?: [:>=, "3.3"], r
           expect(foo_frame[:abs_path]).to include('sentry-ruby/spec/support/profiler.rb')
         end
 
-        it 'has correct stacks', when: { ruby_version?: [:>=, "3.3"], ruby_engine?: "ruby" } do
+        it 'has correct stacks' do
           profile = profiler.to_h[:profile]
           frames = profile[:frames]
           stacks = profile[:stacks]